feat: add dry-run flag

+++ lib/src/assets_list_parser.dart
@@ -109,6 +109,10 @@ class AssetsListParser {
   }
 
   void promptDeletionConfirmation(File asset) {
+    if(config.dryRun) {
+      print('Asset ${asset.path} would be deleted (dry run).');
+      return;
+    }
     if(config.acceptAll) {
       asset.deleteSync();
       print('Asset ${asset.path} has been deleted.');
